+++ RELEASE_NOTES.md
@@ -65,7 +65,7 @@ What's New
 * #1021 - remove .NET 8 support; remove expired deprecations (gbirchmeier)
 * #1015 - rework DateOnlyField/TimeOnlyField to be backed by DateOnly/TimeOnly types instead of DateTime (gbirchmeier)
 * #1023 - new config setting "FieldSeparatorInMessageLogs" (gbirchmeier)
-* new UtcDateTimeField base class so UTCTIMESTAMP fields always have DateTimeKind.Utc (stic)
+* #1030 - new UtcDateTimeField base class so UTCTIMESTAMP fields always have DateTimeKind.Utc (stic)
 
 
 ### v1.14.1